+++ resources/js/emoji.ts
@@ -1,7 +1,6 @@
 import { Controller } from '@hotwired/stimulus';
 import 'emoji-picker-element';
 import { PopupElement } from 'inclusive-elements';
-import twemoji from 'twemoji';
 
 class EmojiPickerController extends Controller<PopupElement> {
     connect() {
@@ -13,32 +12,6 @@ class EmojiPickerController extends Controller<PopupElement> {
         this.element.addEventListener('open', () => {
             pickerRoot.querySelector('input')?.focus();
         });
-
-        const style = document.createElement('style');
-        style.textContent = `.twemoji {
-          width: var(--emoji-size);
-          height: var(--emoji-size);
-          pointer-events: none;
-        }`;
-        pickerRoot.appendChild(style);
-
-        if (Waterhole.twemojiBase) {
-            const observer = new MutationObserver(() => {
-                for (const emoji of pickerRoot.querySelectorAll<HTMLElement>('.emoji')) {
-                    if (!emoji.querySelector('.twemoji')) {
-                        twemoji.parse(emoji, {
-                            base: Waterhole.twemojiBase!,
-                            className: 'twemoji',
-                        });
-                    }
-                }
-            });
-
-            observer.observe(pickerRoot, {
-                subtree: true,
-                childList: true,
-            });
-        }
     }
 }