+++ resources/views/components/post-activity.blade.php
@@ -4,7 +4,9 @@
             {{ Waterhole\username($post->lastComment->user) }}
         </x-waterhole::user-link>
         {{ __('waterhole::forum.post-activity-replied') }}
-        <x-waterhole::time-ago :datetime="$post->last_activity_at" format="micro"/>
+        <a href="{{ $post->urlAtIndex($post->comment_count) }}#bottom" class="color-inherit">
+            <x-waterhole::relative-time :datetime="$post->last_activity_at"/>
+        </a>
     </span>
 @elseif ($post->user)
     <span>
@@ -12,6 +14,8 @@
             {{ Waterhole\username($post->user) }}
         </x-waterhole::user-link>
         {{ __('waterhole::forum.post-activity-posted') }}
-        <x-waterhole::time-ago :datetime="$post->created_at" format="micro"/>
+        <a href="{{ $post->url }}" class="color-inherit">
+            <x-waterhole::relative-time :datetime="$post->created_at"/>
+        </a>
     </span>
 @endif