+++ resources/views/components/user-profile.blade.php
@@ -27,7 +27,7 @@
                     @endif
 
                     @if ($user->bio_html)
-                        <div class="content user-profile__bio">{{ Waterhole\emojify($user->bio_html) }}</div>
+                        <div class="content user-profile__bio">{{ $user->bio_html }}</div>
                     @endif
 
                     <div class="row gap-sm wrap align-center color-muted text-xs user-profile__info">