+++ resources/views/widgets/feed.blade.php
@@ -23,7 +23,7 @@
             >{{ $item->getTitle() }}</a>
 
             <p class="color-muted text-xxs">
-                <x-waterhole::time-ago :datetime="$item->getDateCreated()"/>
+                <x-waterhole::relative-time :datetime="$item->getDateCreated()"/>
                 — {{ Str::limit(htmlspecialchars_decode(strip_tags($item->getDescription() ?: $item->getContent()), 200)) }}
             </p>
         </article>